% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alcohol_2019}
\alias{alcohol_2019}
\title{Alcohol consumption around the world (2019)}
\format{
A data frame with 188 rows and 4 variables:
\describe{
  \item{country}{Country of origin of participant}
  \item{code}{Country code related to the country-of-origin of participant}
  \item{year}{The year in which data was collected}
  \item{consumption}{Indicates the liters of pure alcohol consumed by individuals of 15+ years of age}
}
}
\source{
Multiple sources compiled by World Bank (2024) – processed by Our World in Data.
“Total alcohol consumption per capita – In liters of pure alcohol, projected estimates,
15+ years of age”, Url: \url{https://ourworldindata.org/alcohol-consumption}. World Health Organization (via World Bank),
“World Development Indicators”, Url: \url{https://datacatalog.worldbank.org/search/dataset/0037712/World-Development-Indicators}.
}
\usage{
alcohol_2019
}
\description{
A dataset which contains information about alcohol consumption across various countries.
}
\keyword{datasets}
