% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comments}
\alias{comments}
\title{Student Life (Open comments)}
\format{
A data frame with 209 rows and 1 variable:
\describe{
  \item{comment}{Comments by students about their study experiences.}
}
}
\source{
Data was generated using GPT-3 and GPT-4o by OpenAI. Date of creation: 2024-09-30.
}
\usage{
comments
}
\description{
This dataset contains English comments by students about their study experiences at a fictitious university.
}
\keyword{datasets}
