% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create_dr}
\alias{create_dr}
\title{Automatically create folders for a project}
\usage{
create_dr()
}
\value{

}
\description{
Automatically create folders for a project
}
\details{
This function allows to automatically generate a project
    folder structure in the working directory of your R session.
}
