% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dir_mov}
\alias{dir_mov}
\title{Top 250 movies ranked on IMDb \emph{(Directors with >2 movies)}}
\format{
A data frame with 86 rows and 5 variables:
\describe{
  \item{director}{\emph{Director of the movie}}
  \item{imdb_rating}{\emph{Rating from imdb, ranges from 1-10}}
  \item{year}{\emph{Year the movie was published}}
  \item{title}{\emph{Movie title}}
  \item{movie}{\emph{Grouping ID, i.e. 1 = first movie, 2 = second movie}}
}
}
\source{
IMDb (\url{https://www.imdb.com})
}
\usage{
dir_mov
}
\description{
A subset of the \emph{imdb_top_250} dataset. It only contains movies of directors
who had more than two movies in this data frame Where directors had more than two
movies ranked on IMDb Top 250, a random sample of two movies was chosen.
}
\keyword{datasets}
