% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eqls}
\alias{eqls}
\title{European Quality of Life Survey Time Series, 2007 and 2011}
\format{
A data frame with 79270 rows and 45 variables:
\describe{
  \item{id}{Participant unique ID}
  \item{wave}{Data was collected in several waves, 2007 and 2011. This allows comparing countries over time.}
  \item{country_name}{Name of country participants resides in.}
  \item{iso_code}{Country ISO code}
  \item{marital_status}{Marital status}
  \item{num_of_children}{Item: \emph{Number of children of your own}}
  \item{rural_urban_living}{Item: \emph{The area in which you live is...?}}
  \item{age_category}{Age of respondent}
  \item{education}{Highest level of education completed}
  \item{gender}{Gender of participant}
  \item{employment}{Employment status}
  \item{trust_people}{Item: \emph{Would you say that most people can be trusted?}
                                   | 1 ("You cannot be too careful") to 10 = ("Most people can be trusted").}
  \item{freq_rel_service_use}{Item: \emph{How frequently do you: Attend religious services, apart from weddings, funerals or christenings?}
                                   | 1 ("Every day or almost every day") to 5 = ("Never").}
  \item{freq_internet_personal}{Item: \emph{How frequently do you: Use the Internet other than for work?}
                                   | 1 ("Every day or almost every day") to 5 = ("Never").}
  \item{freq_sports}{Item: \emph{How frequently do you: Take part in sports or physical exercise?}
                                   | 1 ("Every day or almost every day") to 5 = ("Never").}
  \item{freq_social_activities}{Item: \emph{How frequently do you: Participate in social activities of a club, society, or an association?}
                                   | 1 ("Every day or almost every day") to 5 = ("Never").}
  \item{vol_community_service}{Item: \emph{How often unpaid voluntary work in the last 12 months: Community and social services?}
                                   | 1 ("Every week") to 5 = ("Not at all").}
  \item{vol_education_cult_org}{Item: \emph{How often unpaid voluntary work in the last 12 months: Education, cultural, sports or professional organisation?}
                                   | 1 ("Every week") to 5 = ("Not at all").}
  \item{vol_social_charities}{Item: \emph{How often unpaid voluntary work in the last 12 months: Social movements or charities?}
                                   | 1 ("Every week") to 5 = ("Not at all").}
  \item{vol_political_parties}{Item: \emph{How often unpaid voluntary work in the last 12 months: Political parties, trade unions?}
                                   | 1 ("Every week") to 5 = ("Not at all").}
  \item{vol_other_organisation}{Item: \emph{How often unpaid voluntary work in the last 12 months: Other voluntary organisations?
                                   | 1 ("Every week") to 5 = ("Not at all").}}
  \item{volunteering_frequency}{Volunteering frequency | 1 ("Every week") to 5 = ("Not at all").}
  \item{sei_left_out_soc}{Item: \emph{I feel left out of society}.}
  \item{sei_life_is_complicated}{Item: \emph{Life has become so complicated today that I almost can’t find my way}
                                    | 1 ("Strongly agree") to 5 = ("Strongly disagree").}
  \item{sei_not_recognised}{Item: \emph{I feel that the value of what I do is not recognised by others}
                                    | 1 ("Strongly agree") to 5 = ("Strongly disagree").}
  \item{sei_looked_down}{Item: \emph{Some people look down on me because of my job situation or income}
                                    | 1 ("Strongly agree") to 5 = ("Strongly disagree").}
  \item{sei_feel_close_to_people}{Item: \emph{I feel close to people in the area where I live}
                                    | 1 ("Strongly agree") to 5 = ("Strongly disagree").}
  \item{sei}{Item: \emph{Social Exclusion Index}.}
  \item{mwbi_felt_cheerful}{Item: \emph{Which is closest to how you have been feeling over the last 2 weeks? - I have felt cheerful and in good spirits}
                                    | 1 ("All of the time") to 6 = ("At no time").}
  \item{mwbi_felt_calm}{Item: \emph{Which is closest to how you have been feeling over the last 2 weeks? - I have felt calm and relaxed}
                                    | 1 ("All of the time") to 6 = ("At no time").}
  \item{mwbi_felt_active}{Item: \emph{Which is closest to how you have been feeling over the last 2 weeks? - I have felt active and vigorous}
                                    | 1 ("All of the time") to 6 = ("At no time").}
  \item{mwbi_felt_fresh}{Item: \emph{Which is closest to how you have been feeling over the last 2 weeks? - I woke up feeling fresh and rested}
                                    | 1 ("All of the time") to 6 = ("At no time").}
  \item{mwbi_life_is_interesting}{Item: \emph{Which is closest to how you have been feeling over the last 2 weeks? - My daily life has been filled with things that interest me}
                                    | 1 ("All of the time") to 6 = ("At no time").}
  \item{mwbi_felt_tense}{Item: \emph{Which is closest to how you have been feeling over the last 2 weeks? - I have felt particularly tense}
                                    | 1 ("All of the time") to 6 = ("At no time").}
  \item{mwbi_felt_lonely}{Item: \emph{Which is closest to how you have been feeling over the last 2 weeks? - I have felt lonely}
                                    | 1 ("All of the time") to 6 = ("At no time").}
  \item{mwbi_felt_downhearted}{Item: \emph{Which is closest to how you have been feeling over the last 2 weeks? - I have felt downhearted and depressed}
                                    | 1 ("All of the time") to 6 = ("At no time").}
  \item{mwbi}{Item: \emph{WHO-5 mental wellbeing index}.}
  \item{wlbc_too_tired_household}{Item: \emph{I have come home from work too tired to do some of the household jobs which need to be done}
                                    | 1 ("Several times a week") to 6 = ("Less often/rarely").}
  \item{wlbc_fam_responsibilities}{Item: \emph{It has been difficult for me to fulfil my family responsibilities because of the amount of time I spend on the job}
                                    | 1 ("Several times a week") to 6 = ("Less often/rarely").}
  \item{wlbc_cannot_focus_on_work}{Item: \emph{I have found it difficult to concentrate at work because of my family responsibilities}
                                    | 1 ("Several times a week") to 6 = ("Less often/rarely").}
  \item{wlbc}{Item: \emph{Work-life balance conflict}.}
  \item{hrs_work_first_job}{Item: \emph{How many hours do you normally work per week in your main job?} | (80hrs or more).}
  \item{hrs_work_additional_job}{Item: \emph{About how many hours per week did you work in this additional job?} | (60hrs or more).}
  \item{work_hrs}{The total number of working hours.}
}
}
\source{
European Foundation for the Improvement of Living and Working Conditions. (2015).
        European Quality of Life Time Series, 2007 and 2011: Open Access. UK Data Service.
        SN: 7724, DOI: \url{http://doi.org/10.5255/UKDA-SN-7724-1}.
}
\usage{
eqls
}
\description{
The EQLS dataset is a pan-European survey that captures both the objective conditions of citizens'
lives (e.g., employment, income, health) and their subjective perceptions, including life satisfaction
and happiness.
}
\keyword{datasets}
