% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{halloween}
\alias{halloween}
\title{Halloween}
\format{
A data frame with 5982 rows and 3 variables:
\describe{
  \item{expenses}{Expenses related to Halloween candies}
  \item{distance}{Distance covered while 'Trick-or-treating'}
  \item{country}{Participants country of residence}
}
}
\source{
Ripped from the dying hands of the Pumpkin Lord
}
\usage{
halloween
}
\description{
A dataset which contains data about Halloween consumer behaviour across the US,
the UK and Germany.
}
\keyword{datasets}
