% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hie_2021}
\alias{hie_2021}
\title{The Health Index for England (2021)}
\format{
A data frame with 307 rows and 21 variables:
\describe{
  \item{area_name}{The area the data was collected from.}
  \item{region}{Regional categorisation of areas.}
  \item{year}{The year data was collected.}
  \item{activities_in_life_are_worthwhile}{The average score of respondents (16+) on how
                                           meaningful they find their life's activities.}
  \item{early_years_development}{The percentage of 5-year-olds meeting the expected early
                                 learning goals in personal, social, emotional, physical,
                                 communication, language, mathematics, and literacy at the
                                 end of the Early Years Foundation Stage (EYFS).}
  \item{feelings_of_anxiety}{The average anxiety score of respondents (16+) when asked
                             how they felt the day before.}
  \item{happiness}{The average happiness score of respondents (16+) when asked
                             how they felt the day before.}
  \item{healthy_eating}{The percentage of adults (16+) who report eating five or more
                         portions of fruit and vegetables on a 'usual day'.}
  \item{high_blood_pressure}{The percentage of people reporting high blood pressure
                             (hypertension) - not based on GP records but survey responses.}
  \item{household_overcrowding}{The percentage of households considered overcrowded, based
                                on a standard calculation comparing the number of required
                                rooms to the actual number of rooms.}
  \item{life_expectancy}{Period life expectancy at birth is the average number of years a
                         person born in a specific year and area would live if they experienced
                         that area's current age-specific mortality rates throughout their life.}
  \item{life_satisfaction}{The average life satisfaction score of respondents aged 16 and over.}
  \item{low_level_crime}{The number of police-recorded low-level crimes per 1,000 people, with
                         bicycle theft and shoplifting used as indicators of overall rates.}
  \item{mental_health_conditions}{The percentage of people reporting a mental health condition,
                                  such as depression or anxiety- not based on GP records but
                                  survey responses.}
  \item{overweight_and_obesity_in_adults}{The percentage of adults (18+) who are overweight or
                                          obese, based on self-reported height and weight in a survey.}
  \item{overweight_and_obesity_in_children}{The percentage of children in Reception (aged 4-5)
                                            and Year 6 (aged 10-11) who are overweight or obese.}
  \item{physical_activity}{The percentage of adults (19+) reporting physical activity of 150 minutes
                           or more per week.}
  \item{rough_sleeping}{The number of people sleeping outdoors on a single night in October or November,
                        per 100,000 residents.}
  \item{smoking}{The percentage of adults (18+) who currently smoke cigarettes.}
  \item{unemployment}{The percentage of adults (16+) actively seeking work in the past four
                      weeks and available to start within two weeks, relative to the economically
                      active population.}
  \item{young_ppl_edu_empl_appren}{The percentage of 16 and 17-year-olds in sustained education,
                                   employment, or apprenticeships one year after key stage 4 (after
                                   school Year 11).}
}
}
\source{
Office for National Statistics (ONS), released 16 June 2023, ONS website, Health Index Scores, England,
Url: \url{https://www.ons.gov.uk/peoplepopulationandcommunity/healthandsocialcare/healthandwellbeing/datasets/healthindexscoresengland}.
}
\usage{
hie_2021
}
\description{
A selected range of variables from the the Health Index for England dataset, which is
a comprehensive measure of national health, considering outcomes, behaviors, personal
circumstances, and environmental factors. It provides a single health score and allows
detailed breakdowns by location and topic, facilitating comparisons over time. This
dataset only contains data collected in 2021.
}
\keyword{datasets}
