% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ic_training}
\alias{ic_training}
\title{------------------------------------------------------------------------------
COVID-19 DATA}
\format{
A data frame with 300 rows and 3 variables:
\describe{
  \item{test}{Defines whether a test was performed before or after training}
  \item{communication}{Level of confidence in communication with people from diverse cultural backgrounds}
  \item{teamwork}{Level of confidence in working with people from diverse cultural backgrounds}
  \item{leadership}{Level of confidence in leading people from diverse cultural backgrounds}
  \item{communication2}{Factorised 'communication' with levels "yes and "no"}
  \item{teamwork2}{Factorised 'teamwork' with levels "yes and "no"}
  \item{leadership2}{Factorised 'leadership' with levels "yes and "no"}
  \item{communication3}{Factorised 'communication' with levels "high", "medium", and "low"}
  \item{teamwork3}{Factorised 'teamwork' with levels "high", "medium", and "low"}
  \item{leadership3}{Factorised 'leadership' with levels "high", "medium", and "low"}
}
}
\usage{
ic_training
}
\description{
A dataset which contains COVID cases and deaths, vaccination information and public
health social measures (PHSM) for each country covered by the WHO.
}
\keyword{datasets}
