% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_r4np.R
\name{install_r4np}
\alias{install_r4np}
\title{Install R packages relevant for the book 'R for Non-Programmers: A Guide for Social Scientists'.}
\usage{
install_r4np()
}
\description{
Install R packages relevant for the book 'R for Non-Programmers: A Guide for Social Scientists'.
}
\details{
Installs the following R packages at once:
\itemize{
 \item{tidyverse: }{Install the core tidyverse packages.}
 \item{janitor: }{Clean data, e.g. colum names, etc.}
 \item{naniar: }{Inspect and handle missing data.}
 \item{psych: }{Compute all kinds of statistics relevant to Social Sciences.}
 \item{skimr: }{Retrieve descriptive statistics and 'skim' your data.}
 \item{correlation: }{Advanced functionality to perform correlations.}
 \item{rcompanion: }{Additional functions, such as for effect sizes.}
 \item{effectsize: }{Compute and interpret effect sizes.}
 \item{car: }{Interpret regression results}
 \item{FSA: }{NOT SURE}
 \item{jtools: }{Retrieve nicely formatted output from regressions.}
 }
}
\examples{
\dontrun{
# Run this line of code to install all relevant packages
install_r4np()
}
}
