% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quest_time}
\alias{quest_time}
\title{Start and end dates/times of a questionnaire}
\format{
A data frame with 84 rows and 3 variables:
\describe{
  \item{id}{Unique identifier of participant}
  \item{start}{The date and time at which a participant started the questionnaire}
  \item{end}{The date and time at which a participant finished the questionnaire}
}
}
\source{
A random selection of observations from the Global Education Profiler (GEP).
}
\usage{
quest_time
}
\description{
A dataset which contains start and end dates regarding the completion of a questionnaire.
}
\keyword{datasets}
