% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reading_time}
\alias{reading_time}
\title{Student Reading Time}
\format{
A data frame with 209 rows and 1 variable:
\describe{
  \item{gender}{Gender of the participant}
  \item{degree_level}{Degree level of participant}
  \item{min_reading}{Reading time per week measured in minutes}
}
}
\source{
Data was generated using GPT-3 and GPT-4o by OpenAI. Date of creation: 2024-10-01.
}
\usage{
reading_time
}
\description{
This dataset contains English comments by students about their study experiences at a fictitious university.
}
\keyword{datasets}
