% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simpson}
\alias{simpson}
\title{Global Education Profiler \emph{(random sample)}}
\format{
A data frame with 300 rows and 3 variables:
\describe{
  \item{performance}{\emph{Student's change in performance at school.}}
  \item{happiness}{\emph{Participant's change in happiness}}
  \item{group}{\emph{Group a participant belongs to, i.e. Teacher, Student, or Parent}}
}
}
\source{
Dataset created with '\emph{correlation}' package and function \emph{simulate_simpson()}.
}
\usage{
simpson
}
\description{
A dataset containing information about students' \emph{social integration (si)} experience at
their Higher Education Institutions and their \emph{communication skills (cs)} development.
For \emph{cs}, scores of 'actualy experience' and 'importance to me' are available. This
dataset only contains a random sample of 300 observations out of well over 12,000 students
as of August 2021.
}
\keyword{datasets}
