% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swm}
\alias{swm}
\title{Salary, Work experience, and Mindfulness (SWM)}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{salary}{Salary earned (in USD 1000)}
  \item{work_experience}{Years of working experience}
  \item{mindfulness}{Minutes spent on mindfulness practices in a 5-day working week.}
}
}
\source{
Fictitious dataset created for educational purposes.
}
\usage{
swm
}
\description{
This dataset contains information about employees' salary, their working experience and time spend
on mindfulness practices.
}
\keyword{datasets}
