% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wvs}
\alias{wvs}
\title{World Value Survey (Extract)}
\format{
A data frame with 69578 rows and 8 variables:
\describe{
\item{Participant ID}{\emph{Unique identifier for each participant}}
\item{Country name}{\emph{Country name, partially abbreviated}}
\item{Gender}{\emph{Gender of participant}}
\item{YearOfBirth}{\emph{Year of birth of participant}}
\item{Age}{\emph{Age of participant at the time of data collection}}
\item{relationship_status}{\emph{Relationship status of participants}}
\item{Freedom.of.Choice}{Item: \emph{Some people feel they have completely
free choice and control over their lives, while other people feel that
what they do has no real effect on what happens to them. Indicate how much
freedom of choice and control you feel you have over the way your life
turns out | 1 ("no choice at all") to 10 ("a great deal of choice").}}
\item{Satisfaction-with-life}{Item: \emph{All things considered, how satisfied
are you with your life as a whole these days? | 1 (“completely dissatisfied”)
to 10 (“completely satisfied”)}}
}
}
\source{
World Value Survey (\url{https://www.worldvaluessurvey.org})
}
\usage{
wvs
}
\description{
A dataset which contains selected data from the World Value Survey.
This data was obtained in October 2020.
}
\keyword{datasets}
