% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wvs_nona}
\alias{wvs_nona}
\title{World Value Survey \emph{(Extract)}}
\format{
A data frame with 8564 rows and 6 variables:
\describe{
\item{country}{\emph{Country name, partially abbreviated}}
\item{gender}{\emph{Gender of participant}}
\item{age}{\emph{Age of participant at the time of data collection}}
\item{relationship_status}{\emph{Relationship status of participants}}
\item{freedom_of_choice}{Item: \emph{Some people feel they have completely
free choice and control over their lives, while other people feel that
what they do has no real effect on what happens to them. Indicate how much
freedom of choice and control you feel you have over the way your life
turns out | 1 ("no choice at all") to 10 ("a great deal of choice").}}
\item{satisfaction}{Item: \emph{All things considered, how satisfied
are you with your life as a whole these days? | 1 (“completely dissatisfied”)
to 10 (“completely satisfied”)}}
}
}
\source{
World Value Survey (WVS) (\url{https://www.worldvaluessurvey.org})
}
\usage{
wvs_nona
}
\description{
A cleaned version of the \emph{wvs} dataset.
}
\keyword{datasets}
