% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comments}
\alias{comments}
\title{Student Life (Open comments)}
\format{
A data frame with 209 rows and 1 variable:
\describe{
  \item{comment}{Comments by students about their study experiences.}
  \item{social_inclusion}{Provides a score from 1-10 regarding the level of social inclusion
                          experienced by students. | 1 ("I feel lonely") to 10 = ("I feel par of the campus community").}
}
}
\source{
Fictitious dataset created for educational purposes.
}
\usage{
comments
}
\description{
This dataset contains English comments by students about their study experiences at a fictitious university.
}
\keyword{datasets}
