% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gep}
\alias{gep}
\title{Global Education Profiler \emph{(random sample)}}
\format{
A data frame with 300 rows and 15 variables:
\describe{
  \item{age}{\emph{Age of participant}}
  \item{gender}{\emph{Gender of participant}}
  \item{level_of_study}{\emph{Degree level at the time of the research project}}
  \item{si_socialise_with_people_exp}{\emph{Extent to which participants have opportunities to socialise with people from diverse cultural backgrounds}}
  \item{si_supportive_friends_exp}{\emph{Extent to which participants have supportive friends with a diverse cultural background}}
  \item{si_time_socialising_exp}{\emph{Extent to which participants take time to socialise with people from diverse cultural backgrounds}}
  \item{cs_explain_ideas_imp}{\emph{TEXT HERE (importance)}}
  \item{cs_find_clarification_imp}{\emph{Extent to which participants are able to clarify what others mean (importance)}}
  \item{cs_learn_different_styles_imp}{\emph{Extent to which participants learn about other students communication styles (importance)}}
  \item{cs_explain_ideas_exp}{\emph{TEXT HERE (experience)}}
  \item{cs_find_clarification_exp}{\emph{Extent to which participants are able to clarify what others mean (experience)}}
  \item{cs_learn_different_styles_exp}{\emph{Extent to which participants learn about other students communication styles (experience)}}
}
}
\source{
Global Education Profiler (\url{https://warwick.ac.uk/gep})
}
\usage{
gep
}
\description{
A dataset containing information about students' \emph{social integration (si)} experience at
their Higher Education Institutions and their \emph{communication skills (cs)} development.
For \emph{cs}, scores of 'actualy experience' and 'importance to me' are available.
}
\keyword{datasets}
