% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ic_training}
\alias{ic_training}
\title{IC Training Data}
\format{
A data frame with 96 rows and 11 variables:
\describe{
  \item{name}{Random names created with `randomName` package}
  \item{test}{Defines whether a test was performed before or after training}
  \item{communication}{Level of confidence in communication with people from diverse cultural backgrounds}
  \item{teamwork}{Level of confidence in working with people from diverse cultural backgrounds}
  \item{leadership}{Level of confidence in leading people from diverse cultural backgrounds}
  \item{communication2}{Factorised 'communication' with levels "yes and "no"}
  \item{teamwork2}{Factorised 'teamwork' with levels "yes and "no"}
  \item{leadership2}{Factorised 'leadership' with levels "yes and "no"}
  \item{communication3}{Factorised 'communication' with levels "high", "medium", and "low"}
  \item{teamwork3}{Factorised 'teamwork' with levels "high", "medium", and "low"}
  \item{leadership3}{Factorised 'leadership' with levels "high", "medium", and "low"}
}
}
\source{
Research at the University of Warwick (2018)
}
\usage{
ic_training
}
\description{
A dataset which contains real-world pre- and post-training test scores for
Intercultural Communication trainees. Participants names were replaced with random names.
}
\keyword{datasets}
