% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imdb_top_250}
\alias{imdb_top_250}
\title{Top 250 movies ranked on IMDb}
\format{
A tibble with 250 rows and 16 variables:
\describe{
  \item{title}{\emph{Movie title}}
  \item{rank}{\emph{Overall rank on IMDb}}
  \item{year}{\emph{Year the movie was published}}
  \item{runtime_min}{\emph{Runtime of the movie in minutes}}
  \item{imdb_rating}{\emph{Rating from imdb, ranges from 1-10}}
  \item{metascore}{\emph{Metascore which is based on critics' ratings}}
  \item{synopsis}{\emph{A short summary of the movie}}
  \item{director}{\emph{Director of the movie}}
  \item{star_01}{\emph{Actor/actress listed first for this movie}}
  \item{star_02}{\emph{Actor/actress listed second for this movie}}
  \item{star_03}{\emph{Actor/actress listed third for this movie}}
  \item{star_04}{\emph{Actor/actress listed fourth for this movie}}
  \item{votes}{\emph{Number of viewer votes for this movie}}
  \item{gross_in_m}{\emph{Income generated from this movie in million dollars}}
  \item{genre}{\emph{Genres the movie is categorised in}}
  \item{genre_01}{\emph{First genre listed for this movie}}
}
}
\source{
IMDb (\url{https://www.imdb.com})
}
\usage{
imdb_top_250
}
\description{
A dataset containing details about the top 250 movies
ranked on IMDb.
}
\keyword{datasets}
