% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_r4np.R
\name{install_r4np}
\alias{install_r4np}
\title{Install all R packages relevant for the book.}
\usage{
install_r4np()
}
\description{
Install all R packages relevant for the book.
}
\details{
This package installs the following R packages all at once:
\itemize{
 \item{tidyverse: }{A package to install the core tidyverse packages.}
 \item{janitor: }{A package to help clean data, e.g. colum names, etc.}
 \item{naniar: }{A package to inspect and handle missing data.}
 \item{psych: }{A acpakge to compute all kinds of statistics relevant to Social Sciences.}
 \item{skimr: }{A quick way to retrieve descriptive statistics and 'skim' your data.}
 \item{correlation: }{A package with advanced functionality to perform correlations.}
 \item{rcompanion: }{A package with additional functions, such as for effect sizes.}
 \item{effectsize: }{A package that helps to compute and interpret effect sizes.}
 \item{car: }{A package to help interpret regression results}
 \item{FSA: }{NOT SURE}
 \item{jtools: }{A package which helps to retrieve nicely formatted output from regressions.}
 }
}
\examples{
install_r4np
}
