% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simpson}
\alias{simpson}
\title{Simpson's Paradox: A demonstration dataset \emph{(random sample)}}
\format{
A data frame with 300 rows and 3 variables:
\describe{
  \item{performance}{\emph{Student's change in performance at school.}}
  \item{happiness}{\emph{Participant's change in happiness}}
  \item{group}{\emph{Group a participant belongs to, i.e. Teacher, Student, or Parent}}
}
}
\source{
Dataset created with \emph{correlation::simulate_simpson()}
}
\usage{
simpson
}
\description{
A dataset to demonstrate the Simpson's Paradox.
}
\keyword{datasets}
